(chore) move Checkbox into @components/utility
