+++ client/src/components/activities/ActivityForm/ActivityForm.tsx
@@ -1,9 +1,9 @@
+import { Checkbox } from "@/components/utility/Checkbox/Checkbox";
 import type { ModalId } from "@/lib/modal-ids";
 import modalIds from "@/lib/modal-ids";
 import N from "@/lib/theme/components/form.style";
 import DefaultInput from "@/lib/theme/components/input/DefaultInput.style";
 import TagSelector from "@components/tags/TagSelector/TagSelector";
-import { Checkbox } from "@lib/theme/components/Checkbox";
 import type { ActivityWithIds } from "@t/data/activity.types";
 import DateTimePicker from "./DateTimePicker";
 import S from "./style/ActivityForm.style";
