+++ client/src/components/activities/ActivityForm/DateTimePicker.tsx
@@ -1,5 +1,5 @@
+import { Checkbox } from "@/components/utility/Checkbox/Checkbox";
 import DefaultInput from "@/lib/theme/components/input/DefaultInput.style";
-import { Checkbox } from "@lib/theme/components/Checkbox";
 import { FaInfo } from "react-icons/fa";
 import type { DateTimePickerProps } from "./datetime-picker.types";
 import S from "./style/DateTimePicker.style";
